Added support for groups.
A lot of code has been modified to take groups into account, because those
are element-based and stateful. Shape is now called Primitive and the ne
Shape type gathers a Primitive along with the associated groups. That breaks
retrocompatiblity, so I’ll let you chose what to do about that. Alternative
solution is to modify the Shape type by prepending a Vec<GroupName> for each
variant, but that’s boiler plate and a bit overkill (more code would be
impacted though).