Fix obj::parse looping infinitely.
The fix was made possible by introducing the concept of parser
difference, which is simply computing how much of input data was
consumed between two parser states. Cloning a parser is a very
lightweiht operation as parsers are just borrowed input decorated with
some bytes tracking (usize). We can then easily observe if a successful
parser is successful because it just peeked without consuming anything,
which is a good sign of corrupted data.

In a better world, we should have more strict parsing, especially in
parse_geometries, which as a wildcard pattern-matching (that should be
removed and replaced with a strict alternative, IMHO).

Fixes #62